IftU: new baseframe for L2 Shaxthal Sentry Drone.
From now on, new shaxthal baseframes are designed so that they take
advantage of the ~PAL() image functor to keep a single version of each
frame in the filesystem that gets dynamically recolored at runtime.

The base and surface variation palettes are defined in /base-rc.cfg.

(Backported from trunk. The 1.6 functor is ~RC(A=B).)

[Wesnoth-UMC-Dev SVN r4708]