IftU S2, S14: move Faerie Sprite recruitment from S2 to S14
This is mostly for plot reasons. Anlindë's magic is not supposed to be
as strongly faerie as Elynia, and she certainly should not be able to
command beings who are said to antedate the existence of elves.

I was originally going to move this back to S11 (which is where Faerie
Sprite recruitment might have been introduced long before revision 1),
but it makes more sense to postpone this even further to S14, where
the northern forests finally start playing a role in the gameplay.

More player units recruitment adjustments are forthcoming.

[Wesnoth-UMC-Dev SVN r18382]