IftU: use named events to greatly simplify the implementation of the PICK_UP macro
This allows the event handler to remove itself after picking up an item,
so we don't need to pollute the WML variables table with booleans for
every item.

Additionally, the accompanying [remove_item] is now part of the macro
rather than required to be supplied in the [object] code parameter.

[Wesnoth-UMC-Dev SVN r18505]