IftU: replace Qxs Inferno Abyss terrain with mainline Qxua Ethereal Abyss
The terrain was copied by fabi to DM in wesnoth-umc-dev r4411 using the
same string. After DM's inclusion in mainline for 1.7.0, the terrain was
transferred to the core set as Qxua for 1.9.0 by eleazar. The terrain
remains essentially unchanged from IftU's original version, only with
normalized graphics contrast and texture to match all other chasm
terrains.

[Wesnoth-UMC-Dev SVN r19049]