IftU: don't check movement costs when finding vacant tiles in [hidden_unit]
Fixes a fatal bug caused by the fix for the fatal [hidden_unit]
clobbering bug. This new bug has not appeared in IftU so far, but we're
all about future-proofing here.

From AtS r19214:

Zhangor stage 2 and the Blood Cores in E3S10 have infinite movement
costs on all available terrains, which means that
wesnoth.find_vacant_tile() returns nil and the spawned units are left to
die in limbo, never promoted to the gamemap.

At any rate, [unit] isn't very smart about spawning units either. We
could try checking for movement costs first and fallback to ignoring
them if the result is invalid, but it seems too much to do for little
benefit. Border cases like that haven't been found with the previous
broken (clobbering) implementation, so it doesn't really matter in the
end.

[Wesnoth-UMC-Dev SVN r19215]