Disable Elvish Sylph advancement from Elvish Enchantress (fixes #6)
This solution is not without flaws. For example, the Elvish Enchantress
shown in the help system is the mainline Enchantress, which means she's
displayed as being able to advance to Sylph. Our Enchantress only has a
hidden help page and doesn't appear to advance from anything else.

All this is relatively minor compared to the balancing issues caused by
the Sylph.