S11, S12: Remove gold carryover from S11
As Inky pointed out in the forums, it's pointless since the player is
guaranteed to be able to recall a castleful of units in S12 anyway. Any
bonuses from S11 thus go to waste and become irrelevant as the player
starts S14 on a blank slate economy-wise.

Additionally, removed the bonus label from the optional objective in
S11, and S12's starting gold now permits recruiting the most expensive
set of units available at that point (a castleful of Elvish Scouts).