S10y: Fix "failed to auto-store $unit" error messages on prestart
This was caused by the code using unit for a [foreach] loop variable.
Even though prestart events do not have unit targets, the game seems to
attempt to auto-store $unit anywhere it's used. This is an inoffensive
error message, but the console spam can make it harder to identify real
errors in a debugging session.

This patch also changes usage of $item in the same event to use a
different name for consistency.