S18: New scenario revamp
This adds multiple events to the scenario and makes it more of a vague
call-forward to AtS E1 and E3's endings. It also introduces a "taint"
mechanic that forces players to keep their undead units close to Elynia
and Mal Keshar.

It also makes it so the player does not receive OR require income from
this scenario.