S23B: Fix wmlxgettext error due to string concatenation
It seems you can't use #po: before a line that contains a concatenation
of untranslatable and translatable strings unless (presumably) the very
first item in the concatenation is the translatable string you want
annotated.

e.g. this produces an error in wmlxgettext because it presumably
assumes you're trying to annotate "<b>", which is not translatable:

  # po: Foo
  message="<b>"+_"translatable string"+"</b>"